Refactor setup.sh for improved clarity and structure

+++ setup.sh
@@ -56,27 +56,30 @@ mkdir -p ~/.claude
 cp ./claude-settings.json ~/.claude/settings.json
 # Install Fzf
 if [ ! -d "$HOME/.fzf" ]; then
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-~/.fzf/install -y
+	git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+	~/.fzf/install -y
 fi
 # Install dotnet tools to be able to install git credential manager after reboot
-mkdir -p ~/.dotnet/
-wget https://dot.net/v1/dotnet-install.sh -O ~/.dotnet/dotnet-install.sh
-chmod +x ~/.dotnet/dotnet-install.sh
-~/.dotnet/dotnet-install.sh --channel 8.0
-~/.dotnet/dotnet-install.sh --channel 9.0
+if [ ! -f "$HOME/.dotnet/dotnet" ]; then
+	mkdir -p ~/.dotnet/
+	wget https://dot.net/v1/dotnet-install.sh -O ~/.dotnet/dotnet-install.sh
+	chmod +x ~/.dotnet/dotnet-install.sh
+	~/.dotnet/dotnet-install.sh --channel 8.0
+	~/.dotnet/dotnet-install.sh --channel 9.0
+fi
 
 if [ "$ISWSL" = "yes" ]; then
-# install win32yank to share clipboard between neovim and windows 11
-curl -sLo/tmp/win32yank.zip https://github.com/equalsraf/win32yank/releases/download/v0.1.1/win32yank-x64.zip
-unzip -p /tmp/win32yank.zip win32yank.exe > /tmp/win32yank.exe
-chmod +x /tmp/win32yank.exe
-sudo mv /tmp/win32yank.exe /usr/local/bin/win32yank.exe
-
-# to be able to restore (install dependencies) dotnet solutions (projects)
-curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
-wget -qO- https://aka.ms/install-artifacts-credprovider.sh | bash
+	# install win32yank to share clipboard between neovim and windows 11
+	curl -sLo/tmp/win32yank.zip https://github.com/equalsraf/win32yank/releases/download/v0.1.1/win32yank-x64.zip
+	unzip -p /tmp/win32yank.zip win32yank.exe > /tmp/win32yank.exe
+	chmod +x /tmp/win32yank.exe
+	sudo mv /tmp/win32yank.exe /usr/local/bin/win32yank.exe
+	
+	# to be able to restore (install dependencies) dotnet solutions (projects)
+	curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
+	wget -qO- https://aka.ms/install-artifacts-credprovider.sh | bash
 fi
+
 # Install Oh-My-ZSH
 export RUNZSH="no"
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --keep-zshrc -y
@@ -92,29 +95,29 @@ git clone https://github.com/lukechilds/zsh-nvm ~/.oh-my-zsh/custom/plugins/zsh-
 zsh -i -c "~/debian_setup/setup1.sh"
 
 if [ "$ISWSL" = "no" ]; then
-# Install Wezterm
-curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o /usr/share/keyrings/wezterm-fury.gpg
-echo 'deb [signed-by=/usr/share/keyrings/wezterm-fury.gpg] https://apt.fury.io/wez/ * *' | sudo tee /etc/apt/sources.list.d/wezterm.list
-sudo nala update
-sudo nala install wezterm -y
-# Install my wezterm settings
-cp ~/.config/nvim/.wezterm.lua ~/.wezterm.lua
-
-wezterm start --always-new-process --cwd $(pwd) zsh -c "source setup1.sh"
-# setup first restart script
-AUTOSTARTDIR="$HOME/.config/autostart"
-DESKTOPFILE="$AUTOSTARTDIR/after_first_restart.desktop"
-echo "creating temporary autostart file in $AUTOSTARTDIR"
-sudo mkdir $AUTOSTARTDIR
-sudo cp ./after_first_restart.desktop $AUTOSTARTDIR
-sudo sed -i "s|\${cwd}|$(pwd)|g" $DESKTOPFILE
-sudo chmod +x $DESKTOPFILE
-echo "The stage 1 of the setup has completed !"
-echo "Click [Enter] to reboot your machine and continue with the final stage of the setup"
-read A
-
-# reboot to be able to use the new default shell which is ZSH and for nordvpn to work properly
-reboot
+	# Install Wezterm
+	curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o /usr/share/keyrings/wezterm-fury.gpg
+	echo 'deb [signed-by=/usr/share/keyrings/wezterm-fury.gpg] https://apt.fury.io/wez/ * *' | sudo tee /etc/apt/sources.list.d/wezterm.list
+	sudo nala update
+	sudo nala install wezterm -y
+	# Install my wezterm settings
+	cp ~/.config/nvim/.wezterm.lua ~/.wezterm.lua
+	
+	wezterm start --always-new-process --cwd $(pwd) zsh -c "source setup1.sh"
+	# setup first restart script
+	AUTOSTARTDIR="$HOME/.config/autostart"
+	DESKTOPFILE="$AUTOSTARTDIR/after_first_restart.desktop"
+	echo "creating temporary autostart file in $AUTOSTARTDIR"
+	sudo mkdir $AUTOSTARTDIR
+	sudo cp ./after_first_restart.desktop $AUTOSTARTDIR
+	sudo sed -i "s|\${cwd}|$(pwd)|g" $DESKTOPFILE
+	sudo chmod +x $DESKTOPFILE
+	echo "The stage 1 of the setup has completed !"
+	echo "Click [Enter] to reboot your machine and continue with the final stage of the setup"
+	read A
+	
+	# reboot to be able to use the new default shell which is ZSH and for nordvpn to work properly
+	reboot
 fi
